# `--dedupe-options=partial`: avoid quadratic slowdown on extent count